1690: implemented service to store and retrieve policy server files
at the moment this service isn't called anywhere, but we can plug it later
when we know where it will be used.
